requirements.txt
pynacl
cryptography
pytest
pytest-cov
bandit